Set NeedAddress for TExpr.DebugInfo in TExpr.setNeedAddress()
When TExpr.setNeedAddress() is called for a TExpr.DebugInfo, it will try to
call setNeedAddress() for the inner expression, but will not set it for
the expression itself.

This breaks ILEmitter.emit_and_convert_to_address() which then will assume
that the emit(expr) will push the object, not the address, onto the stack.

Setting the NeedAddress flag for the expression itself fixes this.